+++ srcs/print/ft_putnbr_fd.c
@@ -13,6 +13,6 @@ void ft_putnbr_fd(int fd, int n) {
         write(fd, "-", 1);
     }
     if (nb >= 10)
-        ft_putnbr_fd(nb / 10, fd);
+        ft_putnbr_fd((int)nb / 10, fd);
     write(fd, &"0123456789"[nb % 10], 1);
 }
